fix build error for features organizer event
